Minimize memory allocations in lexer
Make it so that string and integer literals' values
are stored directly in the token.
This allows us to get away with only allocating
memory for string literals with a char buffer
(integers' values are simply stored on the stack).
Other tokens just have lexeme pointers to the source.

The reason we can't simply point to the source
for integers and strings is that their values
may not match the raw source code.
For example, a "0xff" in the source evalutes to 255
as a literal.
For strings, it's because of escape sequences.