Make file errors separate from internal/user ones
When a file error occurs, we simply print out the error message
(with an errno message if there's one) without pointing the user
to a Zymux implementation file.

This is because a lot of file errors happen due to a user's fault,
like trying to read a directory instead of a file,
or a path that doesn't exist.

In those cases, we don't want to point the user to an implementation
file uselessly. They also can't be put under user_error(), as that
function expects an error in a *.zmx file.

Use stat() to tell what a path is (file, directory or other).
This is for better reporting on what path we failed to open/read.