Don't hardcode paths in tests
In test_file.c, we have some tests that try to see
if we can successfully open certain paths or not.
The problem is their paths used to be hardcoded with
a slash delimiter.

It's now updated to use delimiters depending on the user's OS.
Currently defaults to slash, unless it's Windows.