Make unclosed open braces error in interpolation
Now if an open brace is never closed in an interpolated string
the lexer will throw an error.

Previously, the lexer simply just went the entire string without
lexing any interpolation if the first one isn't closed.
This is a problem, as the user may have nested interpolation,
but forget to close the first one which results in
nothing interpolating. We'd rather have an explicit error than that.