Create lib directory and place Lukip on it
Created a lib directory flat on the project and placed Lukip inside it.
This was done because Lukip being inside tests/unit caused us
to hardcode which source files we compile in tests/unit.
Now that Lukip is in lib, the Makefile can generically take all
sources in that directory (tests/unit) and compile them.

(Also did some minor refactoring on the Makefile itself).