Change the name of some files
Changed errors.c/.h to report_error.c/.h because the name
errors is similar to error, which I believe is a GNU C library.
Changed zymux_program.c/.h to program.c/.h for a shorter name
(the ZymuxProgram struct also got shortened to ZmxProgram).

The names of the tests for those files were also changed
to match them as well.