Fix parser sometimes getting stuck when panicking
Make parser advance like normal when panicking,
and simply stores the erroroneous token's placement
to go back to it when synchronizing.

Previously, we used to synchronize by disabling the ability to advance
in macros when panicking until we synchronize so that
we start synchronizing at the erroneous token, not many tokens after.
However, in certain situations that can cause the parser to get stuck
in an infinite loop or recursion.