Add strings, conversion, and runtime data types
Fully support string objects in all stages of Zymux,
including interpolated strings.

Also, add support for converting types into a bool,
or a string internally. This is only done for the VM's use as of now.
However, they'll be reused later
when we parse, compile, and interpret the "as" keyword.

Also, add the data type enum in emitter.h to help in implementing
conversion.