Emit/read longer numbers, and use u32 for arrays
Make the emitter write up to 4 byte integers,
and make the number reader read up to that amount as well.

Change the length and capacity members of dynamic arrays to u32.
The reason was both to optimize the sign out, but also
for uniformity with the number emitter and reader.