Add compiling and interpreting from source
Make some helper functions to give us the options to either
get the compiled function from a source,
or to return an unexecuted VM from a source,
or to even fully execute a source string.

These options should be useful when we need to recompile
a file with the bytecode debug option on after a runtime error,
and when executing imports with more VMs later.