Make char buffers not use dynamic array interface
Disconnect char buffers from the dynamic array interface
provided in data_structures.h.

The reason for this change is to disallow us from accidentally
using the functionalities in data_structures.h on a char buffer,
instead of the ones in char_buffer.c/.h because that would cause
the buffer to not be NULL terminated automatically after operations.

Also, refactor data_structures.h a little to be more flexible,
and allow other manually written data structures (like the VM's stack)
to use the new INCREASE_CAP macro instead of manually writing
the increase for each one of them.