Refactor the VM's binary macros.
Change BIN_OP() and BIN_OP_BOOL() from taking a character
for a built in C operation then performing it,
to taking the number or boolean result of the operation itself
which was done on the objects.

This makes its usage a bit more verbose, but it gives us flexibility
on what kind of operation is performed.

For example, it can now be a built in character operation in C,
or an operation like exponents which requires a function.
The previous implementation limited it to character operations,
so the operations that aren't built in and require a function had
to manually be implement.

Also, modify the names of BIN_OP() and BIN_OP_BOOL().