Return new objects when passing same type to "as"
Ensure that "as_{x}" functions return a new object,
even if the passed/converted object is of the same type that
the function converts to. Like a string being passed to as_string().

This convetion is so that mutable objects end up creating a copy,
instead of returning the same object, which may not be what is expected.