Add token creators and fix some lexer/token bugs
Add functions for creating synthetic tokens, which include ones
for integers, floats, and string literals, as well as a default one
for non-literals (ones that don't store union values in them).

Also, fix 2 bugs. One is equal_token() comparing the length of the
position in the source. This is a bug because synthetic tokens
don't have a position in the source yet are considered equal
to the ones that do.
And, fix a bug where append_lexed_string() stored stringVal's length
without accounting for the NUL character in the length of the buffer.