Rename data_structure.h and some of its macros
- Rename data_structure.h to dynamic_array.h.
- Rename a few macros which had generic names to have the "DA_" prefix.

These changes were because the only 2 data structures that would
have their own implementation were just a dynamic array header
and a pair of header and source files for the hash table
(which will be written soon).
This made combining them in a single header not worth it in my opinion.