Add type check before modulo/division by 0 errors
Perform a check on the types of an operation before checking for 0
on the right hand side of a division/modulo.
This is because if the types aren't numbers,
but the right hand side evaluates to bits of 0 in C
it'll print a "can't divide/modulo by 0" message
instead of warning the user that the types used on the operation
are invalid to begin with.