Fix get_node_pos() may index empty array
Fix a minor issue with get_node_pos() where block nodes
get their position by grabbing the position of the first
statement. That's because the statements array of a block
may be empty (just an empty {} block).

The same is not true for strings getting their position by
their first literal, because each string node is guaranteed
to at least have one string literal inside it,
even if it's an empty ("") string.