Improve as_string(), and fix test_compile_expr()
Make as_string() print a function and iterator signature,
so we know an object is wrapped in an iterator, or that
a string belongs to the name of a function.

Also, fix test_compile_expr() tests which broke after
we optimized out OP_FINISH_STRING into not emitting single strings.