Allow disabling default unreachables on switches
Allow enum-exhaustive switch cases to enable or disable a default
unreachable path. This is so that it catches unexpected numbers
exceeding what an enum handles while it's enabled, but also
allows it to throw compiler warnings when it's disabled and
we forget to handle an enum case in that exhaustive switch.