Improve how GC protects objs and interned strings
Add a "layering" system for protection of objects from the GC.
This solves an issue that the boolean approach had, which is when
multiple protections stack on top of each other in one boolean,
then when one of them is removed the boolean is entirely set to false,
thus removing all protection layers at once.

Also, make it so that interned strings are manually appended to
the protected objects pool if protection is on and an interned string
was found. This is because the interned string may not already be on
the protected object pool, and skipping the NEW_OBJ() call
will not append it there, which risks the interned string being GCed.
This is not an issue for interned bools and null because they
last the whole lifetime of the program automatically.