Remove recompilation for runtime errors
Completely get rid of the recompilation optimization,
and instead always keep track of bytecode since the beginning
normally.

The reason for reverting this process is that it doesn't save
enough memory to justify its complexity, and it disallows writing
specific objects in runtime error messages, because by the time
the message is displayed, the object was already deleted to make
space for recompilation.