Make each function have its own optionals
Make it so that (just like closures) each time a function with optionals
is encountered, it creates a new extended functinon
(now called a runtime func for both closures and optional values),
and replaces it with the original static one on the stack.

This ensures that each time the function is encountered it creates
its own new, independant array of optional values loaded at runtime.