add -Werror in make, and do resolver for loop TODO
Added -Werror now in make and removed its TODO comment,
and also fixed the resolver for loop iterable by grabbing
its number after the variable(s) are declared instead of assuming
only one variable is ever declared.

This however had never caused an issue anyways because the iterable
in a for loop can't be accessed to begin with
(and the VM just peeks at the top element to find the iterable).