Make objects inside collection types debug written
Make it so that objects like strings are always written/printed
in the debug way, so for example lists and maps with strings
have their strings wrapped in quotes.

This makes a list of "hi" and "bye" strings displayed as:
	['hi', 'bye']
instead of:
	[hi, bye]
Which makes it clearer that those are strings, not other types.