Refactor and fix some REPL bugs
Fixed a bug where the compiler didn't print bytecode
when debugging bytecode was enabled in the CLI.

Fixed a bug where repl_line() executed the VM even
if the compilation process failed (just added an if check).

Fixed a bug where converting path to absolute in run_file()
happened before actually allocating the source to check
that the file existed, which caused an incorrect error
message to appear.

Made some refactors in the REPL itself as well,
such as moving the check to prevent VM execution while debugging
into the VM itself.